add 后台系统参数增加获取矿工比功能，链上监听系统增加获取NFT_TOKENID功能，抽奖系统增加设置NFT_TOKENID参数。经验道具，装备卷轴，英雄碎片增加NFT_TOKENID参数

+++ gm-domain/src/main/java/com/gm/modules/user/entity/UserEquipmentEntity.java
@@ -35,9 +35,9 @@ public class UserEquipmentEntity implements Serializable {
 	 */
 	private Long userId;
 	/**
-	 * NFTID
+	 * NFT_TOKENID
 	 */
-	private BigDecimal nftId;
+	private Long nftId;
 	/**
 	 * 矿工数
 	 */
